requirements.txt
google-generativeai
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
pypdf
chromadb
sentence-transformers
langchain-huggingface>=0.3.0
streamlit
python-dotenv
